requirements.txt
alembic==1.18.4
blinker==1.6.2
click==8.1.7
colorama==0.4.6
dnspython==2.8.0
email-validator==2.0.0
Flask==2.3.3
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==3.3.2
idna==3.11
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sendgrid==6.10.0
SQLAlchemy==2.0.47
typing_extensions==4.15.0
Werkzeug==2.3.7
WTForms==3.0.1

